detect loader: move to own file
